Removing use of hard coded default values with a constexpr
Moving interpolation routines for elements into the element class

+++ src/meshchecker.h
@@ -53,7 +53,7 @@ class MeshChecker {
   Adcirc::Geometry::Mesh *m_mesh;
 
   static void printFailedLeveeStatus(
-      Adcirc::Geometry::Boundary *bc, int index,
+      Adcirc::Geometry::Boundary *bc, size_t index,
       double minimumCrestElevationOverTopography);
 };
 }  // namespace Utility
